feat(in-app-messenger): move MessagePreviewContent to common place
