+++ src/components/message-preview-content/messages.js
@@ -1,0 +1,12 @@
+// @flow
+import { defineMessages } from 'react-intl';
+
+const messages = defineMessages({
+    previewError: {
+        defaultMessage: `Sorry, we're having trouble showing this image. `,
+        description: 'Error message for preview not loading an image',
+        id: 'be.messageCenter.previewError',
+    },
+});
+
+export default messages;
