+++ src/components/message-preview-content/styles/MessagePreviewContent.scss
