+++ src/components/message-preview-content/styles/PreviewErrorNotification.scss
