+++ src/features/message-center/messages.js
@@ -12,11 +12,6 @@ const messages = defineMessages({
         description: 'Title for all categories',
         id: 'be.messageCenter.all',
     },
-    previewError: {
-        defaultMessage: `Sorry, we're having trouble showing this image. `,
-        description: 'Error message for preview not loading an image',
-        id: 'be.messageCenter.previewError',
-    },
     product: {
         defaultMessage: 'Product',
         description: 'Title for product category',
